Crate profile_traits

This module contains APIs for the profile crate used generically in the rest of Servo. These APIs are here instead of in profile so that these modules won’t have to depend on profile.

Modules

dummy_tracing

Provides API compatible dummies for the tracing-rs APIs we use if tracing is disabled. Hence, nothing will be traced

generic_callback

generic_channel

ipc

mem

APIs for memory profiling.

time

Macros

debug_event

Emit an event at the debug level

debug_span

Constructs a span at the debug level

info_event

Emit an event at the info level

info_span

Constructs a span at the info level

path

An easy way to build a path for a report.

time_profile

Measure the given callback with the time profiler and (if enabled) tracing.

trace_event

Emit an event at the trace level

trace_span

Constructs a span at the trace level