profile_traits

Macro info_span

macro_rules! info_span {
    ($span_name:literal, $($field:tt)*) => { ... };
    ($span_name:literal) => { ... };
}

Constructs a span at the info level

This macro creates a Span for the purpose of instrumenting code to measure the execution time of the span. If the tracing feature (of the crate using this macro) is disabled, then the Span implementation will be replaced with a dummy, that does not record anything.

Attention: This macro requires the user crate to have a tracing feature, which can be used to disable the effects of this macro.