profile_traits::mem

Function perform_memory_report

pub fn perform_memory_report<F: FnOnce(&mut MallocSizeOfOps)>(f: F)

Invoke the provided function after initializing the memory profile tools. The function is expected to call all the desired [MallocSizeOf::size_of] for allocations reachable from the current thread.