Struct profile_traits::mem::Reporter

pub struct Reporter(pub IpcSender<ReporterRequest>);

A memory reporter is capable of measuring some data structure of interest. It's structured as an IPC sender that a ReporterRequest in transmitted over. ReporterRequest objects in turn encapsulate the channel on which the memory profiling information is to be sent.

In many cases, clients construct Reporter objects by creating an IPC sender/receiver pair and registering the receiving end with the router so that messages from the memory profiler end up injected into the client's event loop.

Methods

impl Reporter

fn collect_reports(&self, reports_chan: ReportsChan)

Collect one or more memory reports. Returns true on success, and false on failure.