profile_traits::time

Enum ProfilerCategory

#[repr(u32)]
pub enum ProfilerCategory {

    Painting = 0,
    Layout = 16,
    ImageSaving = 81,
    ScriptSpawnPipeline = 96,
    ScriptConstellationMsg = 97,
    ScriptDevtoolsMsg = 98,
    ScriptDocumentEvent = 99,
    ScriptEvaluate = 101,
    ScriptEvent = 102,
    ScriptFileRead = 103,
    ScriptFontLoading = 104,
    ScriptImageCacheMsg = 105,
    ScriptInputEvent = 106,
    ScriptNetworkEvent = 107,
    ScriptParseHTML = 108,
    ScriptPlannedNavigation = 109,
    ScriptResize = 110,
    ScriptRendering = 111,
    ScriptSetScrollState = 112,
    ScriptSetViewport = 113,
    ScriptTimerEvent = 114,
    ScriptStylesheetLoad = 115,
    ScriptUpdateReplacedElement = 116,
    ScriptWebSocketEvent = 117,
    ScriptWorkerEvent = 118,
    ScriptServiceWorkerEvent = 119,
    ScriptParseXML = 120,
    ScriptEnterFullscreen = 121,
    ScriptExitFullscreen = 122,
    ScriptWorkletEvent = 123,
    ScriptGeolocationEvent = 124,
    ScriptPerformanceEvent = 125,
    ScriptNavigationAndTraversalEvent = 126,
    ScriptPortMessage = 127,
    ScriptWebGPUMsg = 128,
    ScriptDatabaseAccessEvent = 129,
    TimeToFirstPaint = 144,
    TimeToFirstContentfulPaint = 145,
    TimeToInteractive = 146,
    TimeToLargestContentfulPaint = 147,
    IpcReceiver = 148,
    IpcBytesReceiver = 149,

}

Usage sites of variants marked “Rust tracing only” are not visible to rust-analyzer.

Variants

Painting = 0

Paint is rasterising or presenting.

Not associated with a specific URL.

Layout = 16

The script thread is doing layout work.

ImageSaving = 81

ScriptSpawnPipeline = 96

ScriptConstellationMsg = 97

ScriptDevtoolsMsg = 98

ScriptDocumentEvent = 99

ScriptEvaluate = 101

Rust tracing only: the script thread is executing a script. This may include time doing layout or parse work initiated by the script.

ScriptEvent = 102

ScriptFileRead = 103

ScriptFontLoading = 104

ScriptImageCacheMsg = 105

ScriptInputEvent = 106

ScriptNetworkEvent = 107

ScriptParseHTML = 108

The script thread is parsing HTML, rather than doing other work like evaluating scripts or doing layout.

ScriptPlannedNavigation = 109

ScriptResize = 110

ScriptRendering = 111

ScriptSetScrollState = 112

ScriptSetViewport = 113

ScriptTimerEvent = 114

ScriptStylesheetLoad = 115

ScriptUpdateReplacedElement = 116

ScriptWebSocketEvent = 117

ScriptWorkerEvent = 118

ScriptServiceWorkerEvent = 119

ScriptParseXML = 120

The script thread is parsing XML, rather than doing other work like evaluating scripts or doing layout.

ScriptEnterFullscreen = 121

ScriptExitFullscreen = 122

ScriptWorkletEvent = 123

ScriptGeolocationEvent = 124

ScriptPerformanceEvent = 125

ScriptNavigationAndTraversalEvent = 126

ScriptPortMessage = 127

ScriptWebGPUMsg = 128

ScriptDatabaseAccessEvent = 129

TimeToFirstPaint = 144

Web performance metrics.

TimeToFirstContentfulPaint = 145

TimeToInteractive = 146

TimeToLargestContentfulPaint = 147

IpcReceiver = 148

IpcBytesReceiver = 149

Implementations

impl ProfilerCategory

pub const fn variant_name(&self) -> &'static str