profile_traits

Module mem

APIs for memory profiling.

Structs

MemoryReport

A simple memory report

MemoryReportResult

The results produced by the memory reporter.

ProcessReports

A set of reports belonging to a process.

ProfilerChan

Front-end representation of the profiler used to communicate with the profiler.

ProfilerRegistration

A handle that encompasses a registration with the memory profiler. The registration is tied to the lifetime of this type; the memory profiler unregister the reporter when this object is dropped.

Report

A single memory-related measurement.

Reporter

A memory reporter is capable of measuring some data structure of interest. It’s structured as an IPC sender that a ReporterRequest in transmitted over. ReporterRequest objects in turn encapsulate the channel on which the memory profiling information is to be sent.

ReporterRequest

The protocol used to send reporter requests.

ReportsChan

A channel through which memory reports can be sent.

Enums

ProfilerMsg

Messages that can be sent to the memory profiler thread.

ReportKind

The various kinds of memory measurement.

Constants

SEEN_POINTERS 🔒

Traits

OpaqueSender

A trait to abstract away the various kinds of message senders we use.

Functions

perform_memory_report

Invoke the provided function after initializing the memory profile tools. The function is expected to call all the desired [MallocSizeOf::size_of] for allocations reachable from the current thread.