Module profile_traits::mem

APIs for memory profiling.

Structs

ProfilerChan
Front-end representation of the profiler used to communicate with the profiler.

Report
A single memory-related measurement.

Reporter
A memory reporter is capable of measuring some data structure of interest. It's structured as an IPC sender that a ReporterRequest in transmitted over. ReporterRequest objects in turn encapsulate the channel on which the memory profiling information is to be sent.

ReporterRequest
The protocol used to send reporter requests.

ReportsChan
A channel through which memory reports can be sent.

Enums

ProfilerMsg
Messages that can be sent to the memory profiler thread.

ReportKind
The various kinds of memory measurement.

Traits

OpaqueSender
A trait to abstract away the various kinds of message senders we use.